requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.0
exceptiongroup==1.2.0
Flask==3.0.2
frozenlist==1.4.1
google-api-core==2.17.1
google-auth==2.28.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
openai==1.12.0
packaging==23.2
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
sniffio==1.3.0
soundfile==0.12.1
tqdm==4.66.2
twilio==8.13.0
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==3.0.1
yarl==1.9.4
